feat: bump MAX_UNWIND_SHARDS from 64 to 512, widen shard_id u8 → u16
The 64-shard global limit was routinely hit in production, causing silent
stack trace truncation for any process whose binaries exceeded the budget.

- MAX_UNWIND_SHARDS: 64 → 512 (map declaration only, zero BPF verifier impact)
- shard_id type: u8 → u16 across common/ebpf/userspace (supports up to 65535)
- SHARD_NONE sentinel: 0xFF → 0xFFFF
- ExecMapping padding adjusted to maintain repr(C) alignment

Verified: BPF verification time unchanged (188.6ms → 189.2ms), all 16 e2e
tests and 51 unit tests pass.

+++ profile-bee-ebpf/src/lib.rs
@@ -619,7 +619,7 @@ unsafe fn dwarf_unwind_one_frame(state: &mut DwarfUnwindState, proc_info: &ProcI
 
     // Find the mapping that contains current_ip
     let mut found_mapping = false;
-    let mut shard_id: u8 = SHARD_NONE;
+    let mut shard_id: u16 = SHARD_NONE;
     let mut table_count: u32 = 0;
     let mut load_bias: u64 = 0;
 
@@ -780,7 +780,7 @@ unsafe fn dwarf_copy_stack_regs(
 
         // Find the mapping that contains current_ip
         let mut found_mapping = false;
-        let mut shard_id: u8 = SHARD_NONE;
+        let mut shard_id: u16 = SHARD_NONE;
         let mut table_count: u32 = 0;
         let mut load_bias: u64 = 0;
 
@@ -1091,14 +1091,14 @@ unsafe fn try_fp_step(bp: u64) -> Option<(u64, u64)> {
 /// This avoids verifier state explosion that occurs when the two lookups
 /// are separated by typed wrapper code (MapDef::as_ptr() on the inner map).
 #[inline(always)]
-unsafe fn shard_lookup(shard_id: u8, idx: u32) -> Option<UnwindEntry> {
+unsafe fn shard_lookup(shard_id: u16, idx: u32) -> Option<UnwindEntry> {
     let entry: &UnwindEntry = UNWIND_SHARDS.get_value(shard_id as u32, &idx)?;
     Some(*entry)
 }
 
 #[inline(always)]
 unsafe fn binary_search_unwind_entry(
-    shard_id: u8,
+    shard_id: u16,
     table_count: u32,
     relative_pc: u32,
 ) -> Option<UnwindEntry> {
